Move isReInitRequiredto utils and don't make it deprecated for now
