+++ packages/wrapper/src/utils.ts
@@ -4,8 +4,11 @@
  * ------------------------------------------------------------------------------------------ */
 
 import * as vscode from 'vscode';
-import { WebSocketConfigOptions, WebSocketConfigOptionsUrl } from './commonTypes.js';
+import { WebSocketConfigOptions, WebSocketConfigOptionsUrl, WorkerConfigDirect, WorkerConfigOptions } from './commonTypes.js';
 import { CodePlusFileExt, CodePlusUri, CodeResources } from './editorAppBase.js';
+import { EditorAppClassic } from './editorAppClassic.js';
+import { UserConfig } from './userConfig.js';
+import { EditorAppExtended } from './editorAppExtended.js';
 
 export const createUrl = (config: WebSocketConfigOptions | WebSocketConfigOptionsUrl) => {
     let buildUrl = '';
@@ -103,4 +106,44 @@ export const isEqual = (obj1: unknown, obj2: unknown) => {
     } else {
         return obj1 === obj2;
     }
+};
+
+/**
+ * Checks if the app needs to be re-init
+ * @param editorApp
+ * @param userConfig
+ * @param previousUserConfig
+ * @returns
+ */
+export const isReInitRequired = (editorApp: EditorAppClassic | EditorAppExtended, userConfig: UserConfig, previousUserConfig: UserConfig): boolean => {
+    let mustReInit = false;
+    const config = userConfig.wrapperConfig.editorAppConfig;
+    const prevConfig = previousUserConfig.wrapperConfig.editorAppConfig;
+    const prevWorkerOptions = previousUserConfig.languageClientConfig?.options;
+    const currentWorkerOptions = userConfig.languageClientConfig?.options;
+    const prevIsWorker = (prevWorkerOptions?.$type === 'WorkerDirect');
+    const currentIsWorker = (currentWorkerOptions?.$type === 'WorkerDirect');
+    const prevIsWorkerConfig = (prevWorkerOptions?.$type === 'WorkerConfig');
+    const currentIsWorkerConfig = (currentWorkerOptions?.$type === 'WorkerConfig');
+
+    // check if both are configs and the workers are both undefined
+    if (prevIsWorkerConfig && !prevIsWorker && currentIsWorkerConfig && !currentIsWorker) {
+        mustReInit = (prevWorkerOptions as WorkerConfigOptions).url !== (currentWorkerOptions as WorkerConfigOptions).url;
+        // check if both are workers and configs are both undefined
+    } else if (!prevIsWorkerConfig && prevIsWorker && !currentIsWorkerConfig && currentIsWorker) {
+        mustReInit = (prevWorkerOptions as WorkerConfigDirect).worker !== (currentWorkerOptions as WorkerConfigDirect).worker;
+        // previous was worker and current config is not or the other way around
+    } else if (prevIsWorker && currentIsWorkerConfig || prevIsWorkerConfig && currentIsWorker) {
+        mustReInit = true;
+    }
+
+    if (prevConfig.$type !== config.$type) {
+        mustReInit = true;
+    } else if (prevConfig.$type === 'classic' && config.$type === 'classic') {
+        mustReInit = (editorApp as EditorAppClassic).isAppConfigDifferent(prevConfig, config, false) === true;
+    } else if (prevConfig.$type === 'extended' && config.$type === 'extended') {
+        mustReInit = (editorApp as EditorAppExtended).isAppConfigDifferent(prevConfig, config, false) === true;
+    }
+
+    return mustReInit;
 };
